\name{compute.mean.excess.return}
\alias{compute.mean.excess.return}
\title{Compute Market-Timing Premium}
\usage{
  compute.mean.excess.return(R, pos, Rb)
}
\arguments{
  \item{R}{vector of strategy returns}

  \item{pos}{vector of strategy positions}

  \item{Rb}{vector of raw market returns}
}
\description{
  Excess return over long/short position bias adjusted
  market benchmark
}

