\name{compute.momentum}
\alias{compute.momentum}
\title{Equity curve versus moving average (old mom)}
\usage{
  compute.momentum(R,
    equity = equity.index(R, continuous = F, base = 1, na.fill = list(0, 0, 0)),
    lag.n = 0, smooth.n = 1)
}
\description{
  Equity curve versus moving average (old mom)
}

