\name{Cross}
\alias{Cross}
\title{Remove excessive signals
Gives a "1" or true on the day that x crosses above y Otherwise the result is "0".
To find out when x crosses below y, use the formula Cross(y, x)}
\usage{
  Cross(x, y)
}
\description{
  Remove excessive signals Gives a "1" or true on the day
  that x crosses above y Otherwise the result is "0". To
  find out when x crosses below y, use the formula Cross(y,
  x)
}

