% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%crossover\%}
\alias{\%crossover\%}
\title{Crossover operator}
\usage{
x \%crossover\% ...
}
\description{
Crossover occurs (indicates TRUE, otherwise indicates FALSE) when x value
"crosses over" y value - times t where x[t-1] < y[t-1] and x[t] > y[t]
}

