\name{Days.Calendar}
\alias{Days.Calendar}
\title{Construct a calendar of all days}
\source{
  NYSE Holidays: http://www.nyse.com/pdfs/closings.pdf

  NYSE Trading Hours (Saturdays before 1952-09-29):
  http://www.nyse.com/pdfs/historical_trading_hours.pdf
}
\usage{
  Days.Calendar(y = 1800:2015)
}
\description{
  Warning: Saturdays before 1952-09-29 not yet accomodated
}

