% Generated by roxygen2 (4.0.2): do not edit by hand
\name{financial.index}
\alias{financial.index}
\title{Construct indicator class and define identifiers in the context of different data sources}
\usage{
financial.index(primary_id, ..., currency = NA, multiplier = 1,
  tick_size = NULL, identifiers = NULL, type = "financial.index",
  assign_i = TRUE)
}
\arguments{
\item{lookup}{list of data.source objects to look up the ids. Ordered by preference.}
}
\description{
Modeled after FinancialInstrument::
}

