\name{format.stats}
\alias{format.stats}
\title{Some Title}
\usage{
  format.stats(x, select.rows, select.cols, subset = NULL,
    sort = NULL, ascending = NULL, format = "decimal",
    metrics.in.rows = TRUE, digits = 2,
    row.names = rownames(x), col.names = colnames(x))
}
\description{
  Some Title
}

