\name{is.EOW}
\alias{is.EOW}
\title{End of the Week}
\usage{
  is.EOW(x, weekend = T)
}
\arguments{
  \item{x}{Treat Saturday and Sunday as days ending the
  week? If FALSE, week ends on Friday}

  \item{weekend}{Treat Saturday and Sunday as days ending
  the week? If FALSE, week ends on Friday}
}
\description{
  Value is logical, indicating whether the Date supplied is
  at the end of the week
}

