\name{portfolio}
\alias{portfolio}
\title{constructor for creating an portfolio object.
portfolio object represents implementation of a strategy, and hence have the same name.}
\usage{
  portfolio(data)
}
\description{
  constructor for creating an portfolio object. portfolio
  object represents implementation of a strategy, and hence
  have the same name.
}

