\name{Prepare.indicators}
\alias{Prepare.indicators}
\title{Some Title}
\usage{
  Prepare.indicators(x, src = c("package"))
}
\arguments{
  \item{src}{character string indicating where to get the
  indicators data from. If 'package' string is provided,
  then data(,env=indicators)}
}
\description{
  Some Title
}

