% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Search}
\alias{Search}
\title{this is old search instrument function dependent on FinancialInstrument package}
\usage{
Search(pattern, table = "INSTRUMENT", fields = "all")
}
\description{
this is old search instrument function dependent on FinancialInstrument package
}

