\name{showInstruments}
\alias{showInstruments}
\title{blabla}
\usage{
  showInstruments()
}
\arguments{
  \item{x}{An xts object with OHLC-like structure
  (quantmod::is.OHLC(x) == TRUE)}
}
\description{
  Exchange Open and close times This should be
  ticker-specific TODO: Currently Package global variable.
  Rework to be ticker/exchange specific.
}

