\name{sig.compress}
\alias{sig.compress}
\title{x is a data.table Uninformative values are removed: NAs and repetitive values. Only
the leading signal is retained.}
\usage{
  sig.compress(x)
}
\description{
  x is a data.table Uninformative values are removed: NAs
  and repetitive values. Only the leading signal is
  retained.
}

