\name{strategery}
\alias{strategery-package}
\docType{package}

\title{Quant Strategy Specification, Backtesting, Optimization And Statistical
    Analysis Workflow}

\description{
  Bar-driven backtest vs event-driven backtest.
  User defines and analyses strategy in the global environment interactively.
  Handles irregular, asynchronous time series well (indicators and signals).
  Stores data in data.tables, encapsulated in the slots of objects such as indicators
  or signals.
}