\name{Universe}
\alias{Universe}
\title{Define universe of trading opportunities (instruments by bars)}
\usage{
  Universe(..., load.path = NULL)
}
\description{
  Loads INSTRUMENT and OHLCV data for trading. Subsets by
  instruments specified. Constructs trading bars frame for
  each instrument
}

