% Generated by roxygen2 (4.0.2): do not edit by hand
\name{portfolio}
\alias{portfolio}
\title{Keeps track of all positions with a profit and loss ("PnL").}
\usage{
portfolio(name, prices, equity = 0, as.of = min(index(prices)) - 1e-05)
}
\description{
TODO: add benchmarks
TODO: solve contract multipliers in Portfolio$add_txns
TODO: interim on-the-fly valuation (equity value), without saving - for instant exposures
TODO: period summary of "Realized.PL", "Unrealized.PL", "Gross.Trading.PL", "Txn.Fees" if needed
}
\details{
Constraints:
Specifying portfolio constraints as functions of time and asset.
Constraints are applied when determining `sizing` of order in terms of units
p$constraints$max = function(t, symbols) {
   c(DBC=0.002,IEF=0.004)[symbols]
}
p$constraints$min = function(t, symbols) {
  -c(DBC=0.002,IEF=0.004)[symbols]
}
}
\examples{
p = portfolio("ABC", prices=prices, equity=100000)
}

