% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Universe}
\alias{Universe}
\title{Define universe of trading opportunities (instruments by bars)}
\usage{
Universe(..., load.path = file.path(system.file(package = "strategery"),
  "data"))
}
\description{
Loads INSTRUMENT and OHLCV data for trading. Subsets by instruments specified.
Constructs trading bars frame for each instrument
}

